Initial Project Commit
Updated the layout of the project to fall in line with the typical go
project structure of any libraries within the pkg directory and all
commands within the cmd directory.

Also added support for specifying the connection string as a command
line argument and cleaned up some of the output structure.